Rename command module to commands
